Document map props
This mainly documents the `ov5_021E15F4.c` file, which is in charge of
reading map props from the land data NARC, managing them, and rendering
them.

A "map prop" is defined here as a 3D model that is separate from the map
3D model, and that can be positioned, rotated, and scaled independently.
For instance, they are usually used to place buildings on the various
maps.

Signed-off-by: Kuruyia <github@kuruyia.net>