Document persisted map features
This documents the `unk_02027F50.c` file, which is in charge of managing
data used by some maps to be more dynamic, and that needs to be
persisted in the save file. Examples include Sunyshore Gym's bridges
rotation, Pastoria Gym's water level, Canalave Gym's platform
positions...

This also documents related code in the save data module, and names
constants used as IDs for identifying the map that is currently using
dynamic features.

Signed-off-by: Kuruyia <github@kuruyia.net>