Document platform lift-related map script commands
This documents the `ScrCmd_25C` map script command, that is used on maps
containing a platform lift to trigger it going up or down. It is
intended to be used in a script that is called on a coord event.

This also documents the `ScrCmd_25D` map script command, that is used on
the Pokémon League elevator rooms to check if the platform lift has
already been used (i.e. it is on the top floor) when the player entered
the map. This is used to disable the coord event that triggers the
platform lift.

Signed-off-by: Kuruyia <github@kuruyia.net>