Rename constants for window frame tile count
Every type of standard window uses the same number of tiles for their
border, so we can have a single constant instead of one for each type.

This also switches to frame instead of border to align with
`Window_Draw...Frame`.

Finally, this renames the base tile constants in
`wonder_cards_app.c` to align with the `border` -> `frame` change.