Update pokeradar.c for "unk_BC"
Replaced GFXTestBox name "unk_BC" with "grassView".
Suggested due to the GFXTestBox being used to determine if a patch is in the given GFXTestBox; it seems as though any patch that isn't within the bounds that is selected to be active is rendered inactive. Hence the name.
(Apologies if this is inaccurate in any way; I do not wish to make more work. I am trying VERY HARD to suggest a thing that, if it would be an issue, won't BREAK THE BUILD. If my analysis is wrong feel free to let me know.)