Document distortion world floating platform jump points
This continues documentation of the distortion world overlay by
documenting the contents of the second section of files in the
`tw_arc.narc` archive.

This section contains the data for the jump points used by the player to
enter or exit a floating platform. It contains info about where the jump
point is located, where the player will be transported to (both position
and sprite direction), what the destination floating platform is and how
fast the animation is.

The code itself is mostly about animating the player movement from the
jump point to its destination, and animating the player sprite so it
corresponds to the floating platform orientation.

Signed-off-by: Kuruyia <github@kuruyia.net>