Document distortion world ghost props data
This continues documentation of the distortion world overlay by
documenting the contents of the fourth section of files in the
`tw_arc.narc` archive.

This section contains the data for the map props that dynamically appear
and disappear from view when the player is reaching certain triggers.

A trigger is activated when the player is inside the trigger's bounds
and is looking at a certain direction. When activated, the trigger will
show or hide a group of props based on how it is configured.

A ghost prop is part of a group of props which are shown or hidden
together. It also has a kind (a floating platform, a land rock, a
vine-like flower...) that determines the model to render, and a location
where to render that model.

This commit only documents the data related to ghost props found in the
archive. The "runtime" side of ghost props will be documented in a
subsequent commit.

Signed-off-by: Kuruyia <github@kuruyia.net>