Document the overworld animation manager
This documents the `unk_020711EC.c` file, which manages some animations
in the overworld for 2D (e.g. parts of the HM animation) and 3D (e.g.
distortion world's rocks and flowers) elements.

This was supposedly initially used to animate positions, although this
can be (and is being) used to animate other properties as well.

Signed-off-by: Kuruyia <github@kuruyia.net>